chore: table warn + move height function
