+++ crates/common/src/utils.rs
@@ -8,11 +8,13 @@ use anyhow::{anyhow, Context as _};
 use borsh::BorshDeserialize;
 use citrea_evm::system_contracts::{BitcoinLightClientContract, BridgeContract};
 use citrea_evm::{CallMessage as EvmCallMessage, SYSTEM_SIGNER};
+use citrea_primitives::forks::get_forks;
 use reth_primitives::{Recovered, TransactionSigned};
 use sov_db::ledger_db::SharedLedgerOps;
 use sov_modules_api::DaSpec;
 use sov_rollup_interface::rpc::block::L2BlockResponse;
 use sov_rollup_interface::services::da::DaService;
+use sov_rollup_interface::spec::SpecId;
 use sov_rollup_interface::stf::StateDiff;
 use sov_rollup_interface::transaction::Transaction;
 
@@ -136,4 +138,29 @@ pub async fn decode_sov_tx_and_update_short_header_proofs<Da: DaService, DB: Sha
 
 pub fn read_env(key: &str) -> anyhow::Result<String> {
     env::var(key).map_err(|_| anyhow::anyhow!("Env {} missing or invalid UTF-8", key))
+}
+
+// If tangerine activation height is 0, return 1
+// Because in tests when the first l2 block for the first sequencer commitment is needed
+// Tangerine activation height should be sent
+// If it is 0, it errors out because l2 block 0 is not valid
+// So for only in tests, if tangerine activation height is 0, return 1
+// In production, it will return whatever the activation height is
+// If network starts from Fork3, use 1 as the activation height
+// as we'd like to behave the same way
+pub fn get_tangerine_activation_height_non_zero() -> u64 {
+    let forks = get_forks();
+
+    if forks[0].spec_id > SpecId::Tangerine {
+        return 1;
+    }
+
+    let fork = forks
+        .iter()
+        .find(|f| f.spec_id == SpecId::Tangerine)
+        .expect("Tangerine should exist");
+    if fork.activation_height == 0 {
+        return 1;
+    }
+    fork.activation_height
 }
