+++ crates/fullnode/src/da_block_handler.rs
@@ -11,7 +11,8 @@ use anyhow::anyhow;
 use citrea_common::backup::BackupManager;
 use citrea_common::cache::L1BlockCache;
 use citrea_common::da::{extract_zk_proofs_and_sequencer_commitments, sync_l1, ProofOrCommitment};
-use citrea_primitives::forks::{fork_from_block_number, get_tangerine_activation_height_non_zero};
+use citrea_common::utils::get_tangerine_activation_height_non_zero;
+use citrea_primitives::forks::fork_from_block_number;
 use reth_tasks::shutdown::GracefulShutdown;
 use rs_merkle::algorithms::Sha256;
 use rs_merkle::MerkleTree;
