+++ crates/primitives/src/forks.rs
@@ -82,28 +82,3 @@ const _CHECK_FORKS: () = {
         panic!("FORKS order is invalid")
     }
 };
-
-// If tangerine activation height is 0, return 1
-// Because in tests when the first l2 block for the first sequencer commitment is needed
-// Tangerine activation height should be sent
-// If it is 0, it errors out because l2 block 0 is not valid
-// So for only in tests, if tangerine activation height is 0, return 1
-// In production, it will return whatever the activation height is
-// If network starts from Fork3, use 1 as the activation height
-// as we'd like to behave the same way
-pub fn get_tangerine_activation_height_non_zero() -> u64 {
-    let forks = get_forks();
-
-    if forks[0].spec_id > SpecId::Tangerine {
-        return 1;
-    }
-
-    let fork = forks
-        .iter()
-        .find(|f| f.spec_id == SpecId::Tangerine)
-        .expect("Tangerine should exist");
-    if fork.activation_height == 0 {
-        return 1;
-    }
-    fork.activation_height
-}
